Added if statements and pickedUp Checker for kbrd
Basically, I added a variable "pickedUp" to the program. By default, the value of pickedUp is set to 0, which signifies that the passenger has NOT been picked up by the robot, and thus is still on the playing field. If a robot goes over the green zone with a pickedUp value of 0, the kbrdcontrols script will NOT be run (as there's no passenger to drop off). The same can be said for the vice-versa case for the blue zone. However, if the robot goes to a blue zone with a pickedUp value of 0, or a green zone with a pickedUp value of 1, the kbrdcontrols script will run in order for one of us to manually pick up/drop off the robot at the appropriate zone.